update waterRate
зробив документи контролеру та сервісу, також оновив валідацію dailyNormaWaterSchema
перевір будь ласка
