+++ src/routes/index.js
@@ -1,14 +1,17 @@
 import { Router } from 'express';
 import waterRouter from './water.js';
+import waterRateRouter from './water.js';
+
 import authRouter from './auth.js';
 
-import userRouter from './user.js'
+import userRouter from './user.js';
 import monthRouter from './month.js';
-
 
 const router = Router();
 
 router.use('/water', waterRouter);
+router.use('/water-rate', waterRateRouter);
+
 router.use('/auth', authRouter);
 
 router.use('/user', userRouter);
